+++ src/plugins/paging/virtualScroll/virtualScrollPaging.interface.ts
@@ -1,4 +1,4 @@
-import {PagingOptions, Paging} from '../paging.interface';
+import {Paging, PagingOptions} from '../../../interfaces';
 
 /**
  * Css classes for virtual scroll paging
@@ -15,7 +15,7 @@ export interface VirtualScrollPagingOptions extends PagingOptions<CssClassesVirt
     /**
      * Offset between 0 and 1 which indicates when to start loading next page, 0.8 means 80% of scrolling distance
      */
-    loadOffsetTreshold?: number;
+    loadOffsetTreshold: number;
 }
 
 /**